Add files via upload
This is a genetic algorithm I built in class. It uses tournament and roulette wheel selection to get the best fitness. The best fitness would be getting the string of 20 1's or 0's to all 1's, then the algorithm finishes, the user can edit the variables such as initial population, crossover, selection method, and mutation rate if they wish.