Add files via upload
This genetic algorithm builds a schedule based on given classrooms, professors, courses, and timeslots. It uses three objects to construct the chromosome and then there are checks to be sure that this chromosome is a valid schedule. Meaning there are no duplicate courses or multiple classes in the same room at the same time. The user can change the mutation rate, crossover, selection method, and the initial population. The fitness calculation is based on efficient class space usage and consecutive classes in the same room for a single professor.